Renderer: ShaderCompiler code clean-up and performance improvements
# Conflicts:
#	FlyleafLib/MediaFramework/MediaRenderer/ShaderCompiler.cs
